Add files via upload
I am glad to share that I have completed the phase-2 task given by the CodersCave team as a Data Science intern. I was responsible for developing and executing the classification of mobile prices using several libraries like matplotlib and seaborn. I completed the process of testing and training the datasets ,checked for outliers, provided the required information on confusion matrix. Also, done the exploratory data analysis and statistical analysis.